app/requirements.txt
appdirs==1.4.4
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
Flask==1.1.2
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kaleido==0.2.1
kiwisolver==1.3.1
librosa==0.9.2
llvmlite==0.36.0
MarkupSafe==1.1.1
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
plotly==5.13.0
pooch==1.6.0
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.27.1
resampy==0.4.2
retrying==1.3.4
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
soundfile==0.11.0
tenacity==8.1.0
threadpoolctl==3.1.0
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
typing_extensions==4.1.1
urllib3==1.26.14
Werkzeug==2.0.3
WTForms==3.0.0
